apps/ecs-sandbox/pyproject.toml
[project]
name = "ecs-sandbox"
version = "0.1.0"
description = "FastAPI control plane for managing ephemeral sandbox containers"
requires-python = ">=3.12"
dependencies = [
    "ecs-sandbox-client",
    "uvicorn",
    "fastapi",
    "pydantic",
    "aiosqlite",
    "sqlalchemy[asyncio]",
    "docker",
    "redis",
    "taskiq[reload]",
    "taskiq-redis",
    "taskiq-fastapi",
    "python-dotenv",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "black",
    "ruff",
    "ty",
    "httpx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
